Improve inbound message handling and related cleanup

+++ community/bolt/src/main/java/org/neo4j/bolt/v1/messaging/Neo4jPack.java
@@ -69,5 +69,5 @@ public interface Neo4jPack
 
     Unpacker newUnpacker( PackInput input );
 
-    int version();
+    long version();
 }
